#27 module 구성
