+++ app/build.gradle
@@ -60,11 +60,8 @@ dependencies {
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutineVersion"
 
     implementation "com.squareup.retrofit2:retrofit:$retrofitVersion"
-    implementation "com.squareup.retrofit2:converter-gson:$retrofitVersion"
-    implementation "com.squareup.retrofit2:converter-moshi:$retrofitVersion"
 
     // OkHttp Logging
-    def loggingVersion = "4.9.1"
     implementation "com.squareup.okhttp3:logging-interceptor:$loggingVersion"
     implementation "com.squareup.moshi:moshi:$moshVersion"
