--- app/src/main/java/com/mashup/kkyuni/network/NetworkModule.kt
@@ -1,27 +1,0 @@
-package com.mashup.kkyuni.network
-
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import okhttp3.OkHttpClient
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
-import javax.inject.Named
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-object NetworkModule {
-    const val YOUTUBE_API_URL = "https://www.googleapis.com/youtube/v3"
-
-    @Named("youtube_api")
-    @Provides
-    @Singleton
-    fun provideRetrofit(okHttpClient: OkHttpClient): Retrofit =
-        Retrofit.Builder()
-            .baseUrl(YOUTUBE_API_URL)
-            .client(okHttpClient)
-            .addConverterFactory(MoshiConverterFactory.create())
-            .build()
-}
