+++ data/network/build.gradle
@@ -33,24 +33,12 @@ android {
 
 dependencies {
 
-    implementation 'androidx.core:core-ktx:1.6.0'
-    implementation 'androidx.appcompat:appcompat:1.3.1'
-    implementation 'com.google.android.material:material:1.4.0'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
-
-    // Hilt
-    def hilt_version = "2.38.1"
     implementation "com.google.dagger:hilt-android:$rootProject.hiltVersion"
     kapt "com.google.dagger:hilt-android-compiler:$rootProject.hiltVersion"
 
-    // Retrofit
-    def retrofitVersion = "2.9.0"
+    implementation "com.squareup.moshi:moshi:$moshVersion"
+    implementation "com.squareup.retrofit2:converter-moshi:$retrofitVersion"
     implementation "com.squareup.retrofit2:retrofit:$retrofitVersion"
-    implementation "com.squareup.retrofit2:converter-gson:$retrofitVersion"
 
-    // OkHttp Logging
-    def loggingVersion = "4.9.1"
     implementation "com.squareup.okhttp3:logging-interceptor:$loggingVersion"
 }
