+++ feature/calendar/data/build.gradle
@@ -32,14 +32,10 @@ android {
 }
 
 dependencies {
-
     implementation(project(":feature:calendar:domain"))
 
-    implementation 'androidx.core:core-ktx:1.6.0'
+    implementation "com.squareup.retrofit2:retrofit:$retrofitVersion"
+
     implementation "com.google.dagger:hilt-android:$hiltVersion"
     kapt "com.google.dagger:hilt-android-compiler:$hiltVersion"
-
-    testImplementation 'junit:junit:4.+'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 }
