+++ feature/calendar/data/src/main/java/com/mashup/kkyuni/feature/calendar/data/CalendarService.kt
@@ -1,0 +1,5 @@
+package com.mashup.kkyuni.feature.calendar.data
+
+interface CalendarService {
+
+}
