--- feature/calendar/data/src/main/java/com/mashup/kkyuni/feature/calendar/data/di/NetworkModule.kt
@@ -1,12 +1,0 @@
-package com.mashup.kkyuni.feature.calendar.data.di
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@InstallIn(SingletonComponent::class)
-@Module
-class NetworkModule {
-
-
-}
